+++ SharedNote/Resources/GoogleService-Info.plist
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>410670636961-hk692is29mjvcimk2f6hejc0is6o94ij.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.410670636961-hk692is29mjvcimk2f6hejc0is6o94ij</string>
+	<key>API_KEY</key>
+	<string>AIzaSyDYs15hTQ-RKZVDL9QISOO0v_81WoNcyk4</string>
+	<key>GCM_SENDER_ID</key>
+	<string>410670636961</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>com.mrdougwright.SharedNote</string>
+	<key>PROJECT_ID</key>
+	<string>sharednote-taehae</string>
+	<key>STORAGE_BUCKET</key>
+	<string>sharednote-taehae.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false/>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false/>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true/>
+	<key>IS_GCM_ENABLED</key>
+	<true/>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true/>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:410670636961:ios:a3277b2e48dfd98bd235e0</string>
+</dict>
+</plist>
